+++ main.py
@@ -5,11 +5,11 @@ Created on Mon Jan 13 12:00:32 2025
 @author: losat
 """
 
-import Diccionario as Dt
+import DictionaryCore as dcore
 import FilterSortModule as filt
-import EntryFileEditor as entryedt
+import FileEditorModule as filedt
 
-new_dict = Dt.Dictionary("entradas_reducidas.json")
+new_dict = dcore.Dictionary("entradas_reducidas.json")
 my_entries = new_dict.all_entries
 #print(my_entries)
 filt.create_print_entry_table(my_entries)
@@ -21,10 +21,10 @@ filt.filter_entries(my_entries, "month", 3)
 filt.filter_entries(my_entries, "monthly", [10,6])
 filt.filter_entries(my_entries, "date", [21,12,2023])
 
-mi_dict2 = Dt.Dictionary("entrada_simple.json")
+mi_dict2 = dcore.Dictionary("entrada_simple.json")
 # print(mi_dict2.list)
-entryedt.add_entry("entrada_simple.json", "bestellen", 14, 1, 2025)
-entryedt.delete_entry("entrada_simple.json", "jaja")
+filedt.add_entry("entrada_simple.json", "bestellen", 14, 1, 2025)
+filedt.delete_entry("entrada_simple.json", "jaja")
 filt.sort_entries(my_entries, "alphabetical")
 # k2 = sorted(my_entries, key=lambda x: x.date[0])
 # filt.list_entries(k2)
